requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pytest
python-dotenv
pre-commit
alembic
autopep8
pytest_bdd
httpx
pytest-cov